Worked on RosterVC and TeamHomeVC
Very little work done on TeamHomeVC but RosterVC is done with UI for the
most part. I wasn't sure how to code it in a way that would make it
easier to implement the database information but hopefully what I have
is easy to work with. This is currently using the force downcast as I'm
not sure how to use the if-let yet.